Refactoring: move files to appropriate packages
